pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sap-report"
version = "0.1.0"
description = "Desktop report runner for SAP queries"
requires-python = ">=3.11"
dependencies = [
  "pandas",
  "openpyxl",
  "psycopg2-binary",
  "hdbcli",
  "xlwt",
  "python-dotenv",
  "tkcalendar",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
